Move dispatch map hanlding to managed code
Removes a bunch of duplicated logic.
