+++ src/Common/src/Internal/Runtime/TypeManagerHandle.Equality.cs
@@ -1,0 +1,44 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+
+namespace Internal.Runtime
+{
+    public partial struct TypeManagerHandle : IEquatable<TypeManagerHandle>
+    {
+        public static bool operator ==(TypeManagerHandle left, TypeManagerHandle right)
+        {
+            return left._handleValue == right._handleValue;
+        }
+
+        public static bool operator !=(TypeManagerHandle left, TypeManagerHandle right)
+        {
+            return left._handleValue != right._handleValue;
+        }
+
+        public override int GetHashCode()
+        {
+            return _handleValue.GetHashCode();
+        }
+
+        public override bool Equals(object o)
+        {
+            if (!(o is TypeManagerHandle))
+                return false;
+
+            return _handleValue == ((TypeManagerHandle)o)._handleValue;
+        }
+
+        public bool Equals(TypeManagerHandle other)
+        {
+            return _handleValue == other._handleValue;
+        }
+
+        public string LowLevelToString()
+        {
+            return _handleValue.LowLevelToString();
+        }
+    }
+}
