+++ src/Common/src/Internal/Runtime/TypeManagerHandle.cs
@@ -3,12 +3,9 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Diagnostics;
 using System.Runtime.InteropServices;
-using System.Runtime.CompilerServices;
-using Internal.Runtime;
 
-namespace System.Runtime
+namespace Internal.Runtime
 {
     /// <summary>
     /// TypeManagerHandle represents an AOT module in MRT based runtimes.
@@ -20,6 +17,15 @@ namespace System.Runtime
     {
         private IntPtr _handleValue;
 
+        // This is a partial definition of the TypeManager struct which is defined in TypeManager.h
+        [StructLayout(LayoutKind.Sequential)]
+        private struct TypeManager
+        {
+            public IntPtr OsHandle;
+            public IntPtr ReadyToRunHeader;
+            public IntPtr DispatchMap;
+        }
+
         public TypeManagerHandle(IntPtr handleValue)
         {
             _handleValue = handleValue;
@@ -30,39 +36,38 @@ namespace System.Runtime
             return _handleValue;
         }
 
-        public override int GetHashCode()
-        {
-            return _handleValue.GetHashCode();
-        }
-
-        public override bool Equals(object o)
-        {
-            if (!(o is TypeManagerHandle))
-                return false;
-
-            return _handleValue == ((TypeManagerHandle)o)._handleValue;
-        }
-
-        public static bool operator ==(TypeManagerHandle left, TypeManagerHandle right)
-        {
-            return left._handleValue == right._handleValue;
-        }
-
-        public static bool operator !=(TypeManagerHandle left, TypeManagerHandle right)
-        {
-            return left._handleValue != right._handleValue;
-        }
-
-        public bool Equals(TypeManagerHandle other)
-        {
-            return _handleValue == other._handleValue;
-        }
-
         public bool IsNull
         {
             get
             {
                 return _handleValue == IntPtr.Zero;
+            }
+        }
+
+        private unsafe TypeManager* AsTypeManagerPtr
+        {
+            get
+            {
+                unsafe
+                {
+                    return (TypeManager*)(((byte*)(void*)_handleValue) - 1);
+                }
+            }
+        }
+
+        public unsafe IntPtr OsModuleBase
+        {
+            get
+            {
+                return AsTypeManagerPtr->OsHandle;
+            }
+        }
+
+        public unsafe IntPtr DispatchMap
+        {
+            get
+            {
+                return AsTypeManagerPtr->DispatchMap;
             }
         }
     }
