+++ src/Common/src/TypeSystem/NativeFormat/NativeFormatMetadataUnit.cs
@@ -6,13 +6,13 @@
 using System;
 using System.Collections.Generic;
 using System.Reflection;
-using System.Runtime;
+
 using Internal.Metadata.NativeFormat;
 using Internal.Runtime.Augments;
-
 using Internal.TypeSystem;
 using Internal.Reflection.Execution;
 using Internal.Reflection.Core;
+using Internal.Runtime;
 using Internal.Runtime.TypeLoader;
 
 using AssemblyFlags = Internal.Metadata.NativeFormat.AssemblyFlags;
