+++ src/Common/src/TypeSystem/NativeFormat/NativeFormatMethod.cs
@@ -7,9 +7,10 @@ using System;
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Reflection;
-using System.Runtime;
 using System.Threading;
+
 using Internal.Metadata.NativeFormat;
+using Internal.Runtime;
 using Internal.Runtime.CompilerServices;
 
 using Internal.TypeSystem;
