+++ src/Runtime.Base/src/System/Runtime/InternalCalls.cs
@@ -223,16 +223,6 @@ namespace System.Runtime
         [ManuallyManaged(GcPollPolicy.Never)]
         internal extern static unsafe bool RhpEHEnumNext(void* pEHEnum, void* pEHClause);
 
-        [RuntimeImport(Redhawk.BaseName, "RhpHasDispatchMap")]
-        [MethodImpl(MethodImplOptions.InternalCall)]
-        [ManuallyManaged(GcPollPolicy.Never)]
-        internal extern static unsafe bool RhpHasDispatchMap(EEType* pEETypen);
-
-        [RuntimeImport(Redhawk.BaseName, "RhpGetDispatchMap")]
-        [MethodImpl(MethodImplOptions.InternalCall)]
-        [ManuallyManaged(GcPollPolicy.Never)]
-        internal extern static unsafe DispatchResolve.DispatchMap* RhpGetDispatchMap(EEType* pEEType);
-
         [RuntimeImport(Redhawk.BaseName, "RhpGetSealedVirtualSlot")]
         [MethodImpl(MethodImplOptions.InternalCall)]
         [ManuallyManaged(GcPollPolicy.Never)]
